+++ docker/zap-webswing.sh
@@ -59,6 +59,27 @@ case "$1" in
         # Run Webswing server- expects X Server to be running
         # dont put into the background otherwise docker will exit
         cd $HOME
+        
+        # Set up the ZAP runtime options
+        ZAP_OPTS="-host 0.0.0.0 -port 8090"
+        ZAP_PUBLIC="/zap/wrk/owasp_zap_root_ca.cer"
+        ZAP_PRIVATE="/zap/wrk/owasp_zap_root_ca.key"
+
+        if [ ! -z "${ZAP_WEBSWING_OPTS}" ]; then
+          # Replace them with those set in the env var
+          ZAP_OPTS="${ZAP_WEBSWING_OPTS}"
+        elif [ -f ${ZAP_PRIVATE} ]; then
+          # Private cert is available, use that
+          ZAP_OPTS="${ZAP_OPTS} -certload ${ZAP_PRIVATE}"
+        elif [ -w /zap/wrk ]; then
+          # wrk directory is writable, output public and private certs
+          ZAP_OPTS="${ZAP_OPTS} -certpubdump  ${ZAP_PUBLIC} -certfulldump  ${ZAP_PRIVATE}"
+        fi
+        
+        echo "Using ZAP command line options: ${ZAP_OPTS}"
+        # Use ; for sed separators so we can use the directory slashes
+        sed -i "s;ZAP_OPTS;${ZAP_OPTS};" webswing.config
+        
         $JAVA_HOME/bin/java $JAVA_OPTS -jar webswing-server.war $OPTS 2>> $LOG >> $LOG
         ;;
     *)