+++ zap/src/main/java/org/parosproxy/paros/core/scanner/ScannerParam.java
@@ -50,6 +50,7 @@
 // ZAP: 2019/06/01 Normalise line endings.
 // ZAP: 2019/06/05 Normalise format/style.
 // ZAP: 2020/11/26 Use Log4j 2 classes for logging.
+// ZAP: 2021/04/13 Issue 6469: Add option to scan null JSON values.
 package org.parosproxy.paros.core.scanner;
 
 import java.util.ArrayList;
@@ -116,6 +117,13 @@ public class ScannerParam extends AbstractParam {
      * @see #addQueryParam
      */
     private static final String SCAN_ADD_QUERY_PARAM = ACTIVE_SCAN_BASE_KEY + ".addQueryParam";
+
+    /**
+     * Configuration key to write/read the {@link #scanNullJsonValues} flag.
+     *
+     * @since 2.11.0
+     */
+    static final String SCAN_NULL_JSON_VALUES = ACTIVE_SCAN_BASE_KEY + ".scanNullJsonValues";
 
     // ZAP: Configuration constants
     public static final int TARGET_QUERYSTRING = 1;
@@ -186,6 +194,17 @@ public class ScannerParam extends AbstractParam {
      */
     private boolean addQueryParam;
 
+    /**
+     * Flag that indicates if the active scanner should scan null JSON values.
+     *
+     * <p>Default value is {@code false}.
+     *
+     * @since 2.11.0
+     * @see #isScanNullJsonValues()
+     * @see #setScanNullJsonValues(boolean)
+     */
+    private boolean scanNullJsonValues;
+
     // ZAP: Excluded Parameters
     private final List<ScannerParamFilter> excludedParams = new ArrayList<>();
     private final Map<Integer, List<ScannerParamFilter>> excludedParamsMap = new HashMap<>();
@@ -240,6 +259,8 @@ public class ScannerParam extends AbstractParam {
         this.scanHeadersAllRequests = getBoolean(SCAN_HEADERS_ALL_REQUESTS, false);
 
         this.addQueryParam = getBoolean(SCAN_ADD_QUERY_PARAM, false);
+
+        this.scanNullJsonValues = getBoolean(SCAN_NULL_JSON_VALUES, false);
 
         // Parse the parameters that need to be excluded
         // ------------------------------------------------
@@ -586,5 +607,29 @@ public class ScannerParam extends AbstractParam {
     public void setAddQueryParam(boolean addQueryParam) {
         this.addQueryParam = addQueryParam;
         getConfig().setProperty(SCAN_ADD_QUERY_PARAM, this.addQueryParam);
+    }
+
+    /**
+     * Tells whether or not the active scanner should scan null JSON values.
+     *
+     * @return {@code true} if null values should be scanned, {@code false} otherwise.
+     * @since 2.11.0
+     * @see #setScanNullJsonValues(boolean)
+     */
+    public boolean isScanNullJsonValues() {
+        return scanNullJsonValues;
+    }
+
+    /**
+     * Sets whether or not the active scanner should scan null JSON values.
+     *
+     * @param scan {@code true} if null values should be scanned, {@code false} otherwise.
+     * @since 2.11.0
+     * @see #isScanNullJsonValues()
+     * @see VariantJSONQuery#setScanNullValues(boolean)
+     */
+    public void setScanNullJsonValues(boolean scan) {
+        this.scanNullJsonValues = scan;
+        getConfig().setProperty(SCAN_NULL_JSON_VALUES, this.scanNullJsonValues);
     }
 }