+++ docker/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to the docker containers will be documented in this file.
 
+### 2021-05-05
+ - Updated to use Webswing 21.1
+
 ### 2021-04-30
  - Alert_on_Unexpected_Content_Types.js > Added Content-Type text/yaml to the list of expected types.